pyproject.toml
[project]
name = "ferret-io"
version = "0.1.0"
description = "Async-native, persistent, inter-process performance logger."
readme = "README.md"
authors = [
    { name = "Alejandro Piad", email = "apiad@apiad.net" }
]
requires-python = ">=3.13"
dependencies = [
    "beaver-db>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["ferret"]

[project.scripts]
ferret = "ferret.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.11.0",
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
